W-12637937: Code  review
